Update investment of a customer and Exception Handling for Delete Customer

+++ maturity-alert/src/main/java/com/pauls/maturity_alert/service/CustomerServiceImpl.java
@@ -1,7 +1,9 @@
 package com.pauls.maturity_alert.service;
 
 import com.pauls.maturity_alert.model.CustomerDetails;
+import com.pauls.maturity_alert.model.InvestmentDetails;
 import com.pauls.maturity_alert.repository.CustomerRepository;
+import com.pauls.maturity_alert.repository.InvestmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +17,9 @@ public class CustomerServiceImpl implements CustomerService{
     @Autowired
     private CustomerRepository customerRepository;
 
+    @Autowired
+    private InvestmentRepository investmentRepository;
+
 
     @Override
     public void addNewCustomer(CustomerDetails customerDetails) {
@@ -24,6 +29,13 @@ public class CustomerServiceImpl implements CustomerService{
 
     @Override
     public void deleteCustomerById(long id) {
+
+        List<InvestmentDetails> investments =
+                investmentRepository.getAllInvestmentsByCustId(id);
+        if(!investments.isEmpty()) {
+            throw new RuntimeException("Cannot delete customer with active investments");
+        }
+
         customerRepository.deleteById(id);
     }
